backend/requirements.txt
flask
flask-cors
pandas
joblib
scikit-learn
mysql-connector-python
numpy